rename talk.models -> models.talk
