+++ backend/backoffice/models/talk.py
