+++ backend/backoffice/tests/models/test_talk.py
@@ -10,7 +10,7 @@ from django.test import TestCase
 from django.utils import timezone
 from django.utils.text import slugify
 
-from backoffice.talk.models import TalkEntry
+from backoffice.models.talk import TalkEntry
 from tournesol.tests.utils.mock_now import MockNow
